user_service/requirements.txt
alembic
asyncpg
fastapi==0.115.8
httpx
motor==3.7.0
passlib[bcrypt]
psycopg2-binary
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose[cryptography]
python-multipart
redis
requests==2.32.3
slowapi
sqlalchemy
typing_extensions==4.12.2
uvicorn==0.34.0